refactor : StringParser에서 FormulaParser로 Class이름 변경
